backend/pyproject.toml
[project]
name = "english-train-backend"
version = "0.1.0"
description = "Backend API for English learning feedback tool"
requires-python = ">=3.10"
dependencies = [
    "fastapi==0.111.0",
    "uvicorn[standard]==0.30.1",
    "python-multipart==0.0.9",
    "pydantic==2.8.2",
    "pydantic-settings==2.3.4",
    "openai-whisper==20231117",
    "torch>=2.0.0",
    "torchaudio>=2.0.0",
    "pydub==0.25.1",
    "ollama==0.3.0",
    "numpy>=1.24.0",
    "ffmpeg-python==0.2.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
only-include = [
    "app.py",
    "stt.py",
    "llm.py",
    "convert.py",
    "models.py",
    "__init__.py",
]

[tool.uv]
dev-dependencies = []
